Refactored routing: separated home and dashboard components into distinct files
