+++ src/pages/dashboard.tsx
@@ -1,0 +1,19 @@
+import { Head } from '@/components/layout/Head'
+import { UserInfo } from '@/components/views/UserInfo'
+import { Statistics } from '@/components/views/Statistics'
+import { LatestBlocksSP } from '@/components/views/LatestBlocksSP'
+import { LatestDonationsSP } from '@/components/views/LatestDonationsSP'
+
+export default function Dashboard() {
+    return (
+        <>
+            <Head />
+            <main>
+                <Statistics />
+                <UserInfo />
+                <LatestBlocksSP />
+                <LatestDonationsSP />
+            </main>
+        </>
+    )
+}
